{"schema":"decodex/commit/1","summary":"Split MCP lane control server modules","authority":"XY-1115"}
